backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
openai==1.31.2
pypdf==4.2.0
python-dotenv==1.0.1
transformers==4.40.2
